M1192: security — seccomp-BPF KILL action (the hard-sandbox completion of M1190)
M1190's filter could only deny a forbidden syscall (return -1); the action real
sandboxes (Chrome/Docker) rely on is KILL — terminate a process that attempts a
forbidden syscall. This adds it.

The filter's RET value is now a 3-way verdict: SECCOMP_RET_DENY(0) -> the
syscall returns -1; SECCOMP_RET_KILL(2) -> the process is killed
(app_sys_exit(-1), mirroring the pledge-violation path, with a dmesg line);
anything else -> ALLOW. So M1190's RET 1=allow / RET 0=deny programs are
byte-compatible and only the new RET 2 kills. app_seccomp_filter_check now
returns the raw verdict (was a 0/1 bool); the gated syscall_dispatch hook
interprets it. The three verdict constants live in the shared syscall.h. No new
syscalls; no ABI change (still the M1190 per-process seccomp_prog).

Verified: make check 55 green (verdict change doesn't regress unfiltered apps or
the M1190 deny path) + headless in-guest — seccomptest now runs both parts: a
DENY child (blocked writefile -> -1, then allowed getpid -> exit 42) AND a KILL
child (KILL-on-writefile filter, calls writefile, terminated before reaching
exit(7) — waitpid confirms it never hit 7): "seccomp: DENY blocks (-1) + allowed
runs; KILL terminates the violator -- OK", shell unaffected.

seccomp-BPF is now a complete sandbox primitive (allow / deny / kill).
Follow-ons: 64-bit arg fields, a TRAP-to-supervisor action bridging to M1124
seccomp-notify, /proc/<pid> filter introspection.

+++ kernel/include/app.h
@@ -124,7 +124,7 @@ int  app_mkfifo(const char *path);     /* create a named pipe (FIFO); 0/-1 (M118
 int  app_fifo_open(const char *path, int write);  /* open a FIFO end -> fd; -1 (M1188) */
 int  app_seccomp_filter_install(const void *prog, int n);   /* install a self-imposed BPF syscall filter; 0/-1 (M1190) */
 int  app_seccomp_filter_active(app_t *a);                   /* does this app have a seccomp-BPF filter? (M1190) */
-int  app_seccomp_filter_check(app_t *a, uint64_t nr, uint64_t a0, uint64_t a1, uint64_t a2);  /* 1 allow / 0 deny (M1190) */
+long app_seccomp_filter_check(app_t *a, uint64_t nr, uint64_t a0, uint64_t a1, uint64_t a2);  /* raw verdict: 0 deny / 2 kill / else allow (M1190, M1192) */
 int  app_deliver_pending(struct registers *r);     /* deliver a pending async signal on return to ring 3; 1 if delivered */
 void app_set_alarm(uint64_t ticks);                /* SYS_alarm: arm a periodic SIGALRM every `ticks` ticks (0=disarm) */
 void app_alarm_tick(void);                         /* timer IRQ hook: raise SIGALRM if the current app's alarm is due */
